refactor(architecture): reorganize diagram feature structure for scalability
- refactor(structure): group layout components (Editor, Canvas, Header, Sidebar) in 'features/diagram/components/layout'.
- refactor(structure): move modals (ClassEditor, Spotlight) to 'features/diagram/components/modals'.
- refactor(structure): move ui atoms (ContextMenu, SidebarItem) to 'features/diagram/components/ui'.
- refactor(nodes): isolate UML specific nodes in 'features/diagram/components/nodes/uml' to allow future diagram types.
- refactor(hooks): promote 'useThemeSystem' to global 'src/hooks' scope as it affects the entire app.
