+++ frontend/src/features/diagram/components/layout/Sidebar.tsx
@@ -14,9 +14,9 @@ import {
   ChevronDown,
   ChevronRight,
 } from "lucide-react";
-import { useDiagramStore } from "../../../store/diagramStore";
-import type { stereotype, UmlRelationType } from "../../../types/diagram.types";
-import { edgeConfig } from "../../../config/theme.config";
+import { useDiagramStore } from "../../../../store/diagramStore";
+import type { stereotype, UmlRelationType } from "../../../../types/diagram.types";
+import { edgeConfig } from "../../../../config/theme.config";
 import { useTranslation } from "react-i18next";
 
 export default function Sidebar() {
