+++ frontend/src/features/diagram/components/nodes/uml/UmlClassNode.tsx
@@ -1,8 +1,8 @@
 import { memo, useState, useEffect } from "react";
 import { Handle, Position, type NodeProps } from "reactflow";
-import type { UmlClassData } from "../../../../types/diagram.types";
-import { useDiagramStore } from "../../../../store/diagramStore";
-import { handleConfig } from "../../../../config/theme.config";
+import type { UmlClassData } from "../../../../../types/diagram.types";
+import { useDiagramStore } from "../../../../../store/diagramStore";
+import { handleConfig } from "../../../../../config/theme.config";
 
 const UmlClassNode = ({ id, data }: NodeProps<UmlClassData>) => {
   const updateNodeData = useDiagramStore((s) => s.updateNodeData);
